Add sidebar styling
This design isn't particularly responsive, elements will get cut
off on the right if the page is narrow. Per the spec, I'm not
worrying too much about responsive design.

For accessibility, sidebar links are in lists. For display, they're
in a grid layout, which means the images and the labels are in
separate list elements. That's okay, because the icons should be
hidden from screenreaders anyway, since the information they
convey is duplicative with the text. (Ref
https://www.a11y-101.com/development/icons-and-links)

If these were actually links, both the images and the text need
to be linkified, and then there'd be a weird non-link gap in the
middle for the gutter... so there might be a better way to do this
than with grid. But hey, this is a project that in part is about
using grid! So I'll stick with this for now.