Style top row of navbar
This change adds a grid layout for the navbar (making it one row
in the layout grid, rather than two) and positions the different
elements within that grid. It also adds styling for the elements
in the top row of the new navbar grid: children are now flex
layouts and sized and centered more-or-less correctly.