Add basic card grid and styling
I chose to add the cards programmatically in javascript mostly
because I needed more practice with javascript dom manipulation
than CSS and HTML, but I should move most of the styling back into
the CSS and just create named classes in the JS.

There's still lots of styling that needs to happen on the cards,
including spacing on the icons and a cut on the height of a card's
body text.

I tried to programmatically alter the SVG color but couldn't get it
working, so manually changed the fill color. Feels skeevier doing
that when it's changing it to grey and not just white!